+++ packages/blade/src/components/ActionList/docs/propsTable.tsx
@@ -4,7 +4,15 @@ import { Code } from '~components/Typography';
 import { ScrollLink } from '~utils/storybook/ScrollLink';
 
 type PropsTableType<T> = Record<
-  keyof Omit<T, 'testID' | '_hideDivider' | '_sectionChildValues' | '_index'>,
+  keyof Omit<
+    T,
+    | 'testID'
+    | '_hideDivider'
+    | '_sectionChildValues'
+    | '_index'
+    | '_virtualizedIndex'
+    | '_onVirtualizedFocus'
+  >,
   string | React.ReactElement | { note: string; type: string | React.ReactElement }
 >;
 const actionListPropsTables: {
@@ -20,8 +28,7 @@ const actionListPropsTables: {
       </>
     ),
     isVirtualized: {
-      note:
-        'Currently only works in ActionList with static height items (items without description) and when ActionList has more than 10 items',
+      note: 'Currently only works with ActionList and ActionListSection',
       type: 'boolean',
     },
   },