+++ packages/blade/src/components/BaseMotion/Motion.test.stories.tsx
@@ -179,6 +179,7 @@ OnlyOutAnimation.play = async ({ canvasElement }) => {
   await expect(getByTestId('card').style.transform).toBe('translateY(0%)');
 
   await userEvent.click(getByRole('button', { name: 'Toggle Animation' }));
+  await sleep(10);
   await waitFor(() => expect(getByTestId('badge')).toHaveStyle('opacity: 0'));
   await waitFor(() => expect(getByTestId('button').style.transform).toBe('translateY(16px)'));
   await waitFor(() => expect(getByTestId('card').style.transform).toBe('translateY(100vh)'));