+++ packages/blade/src/components/DatePicker/FilterChipDatePicker/FilterChipDatePicker.web.tsx
@@ -1,0 +1,8 @@
+import type { FilterChipDatePickerProps } from '../types';
+import { BaseDatePicker } from '../BaseDatePicker.web';
+
+const FilterChipDatePicker = (props: FilterChipDatePickerProps): React.ReactElement => {
+  return <BaseDatePicker {...props} inputElementType="chip" />;
+};
+
+export { FilterChipDatePicker };